studio/requirements.txt
pywebview>=5.0
complier
httpx
